Fix a bug causing scratch dir not to be cleaned up
The desired behavior is that the scratch dir and template dir are both
removed unless the user specifies --keep-temp-dirs. We failed to
actually clean up the scratch dir.

There was also an incorrect test where this was supposed to be tested;
we incorrectly asserted the oppposite of what we should have. This has
been fixed here.