More detailed logic for when overwriting is OK
This improves error messages and defines in detail when files in the
scratch directory can be overwritten. It forbids overwriting files with
directories and vice versa, because that would be confusing for both
template developers and abc code maintainers.

The "overwrite" argument is currently always "true", because we only use
it to write to the scratch directory, but an upcoming PR will use
overwrite=false when writing to the final output directory (technically,
it will use the value of --force-overwrite, which defaults to false).