Implement the "include" and "print" actions (#10)
The "include" action copies files from the template into the scratch
directory. This is analogous to a docker "COPY".

The "print" command just prints to stdout.

In both of these actions, template variable substitution is allowed.

This PR is big, but it's mostly tests.

Coming next: implementing more action types and committing the scratch
dir to the output directory.