More detailed logic for when overwriting is OK (#11)
The current logic is that we will use `overwrite=true` when writing to
the scratch directory. This allows multiple include actions to touch the
same set of files, if the template author wishes.

This PR also improves error messages in some edge cases.

This PR also forbids overwriting files with directories and vice versa,
because that would be confusing for both template developers and abc
code maintainers.

In an upcoming PR, we will set
`overwrite=$(value_of_--force_overwrite_flag)` when writing to the final
output directory.