Add the stringreplace action
This is an action that performs a straightforward string replacement on
a list of paths (either directories or files). Template variables can be
used when specifying any of the parameters:

- paths in which to replace
- strings to be replaced
- strings to be used as replacements.

Most of the logic is in the walkAndModify() function, and the string
replacement logic is a thin wrapper around it. This lays the foundation
for actionRegexReplace() and actionGoTemplate() to also use
walkAndModify().