Copy the scratch dir contents to the dest dir (#13)
This is the final phase of the the rendering process: taking each file
and directory in the scratch dir and copying it into the destination.

We want to make this as atomic as possible. If there are any errors, we
want to leave the destination directory untouched. We don't want half of
the output files to succeed and half of them to be missing; this would
be annoying for users to clean up. To achieve this atomicity, we first
do a "dry run" copy that checks whether the copy *would* succeed, then
we do a real copy. This could be made more efficient in the future, but
it's good enough for now.

In this PR we also switch from ReadFile+WriteFile to io.Copy; it seemed
silly to read a buffer then just write it again.