feat: `include` action can relocate files
Before this, files in the template could only be located in the same
place in the output directory as they were in the template directory.
This is too inflexible. Users will want to do things like "use either
config_a.txt or config_b.txt as the output config.txt depending on some
input variable."

There are two different ways to do this:

Using `as`:

```
  # Copies in_dir_1/file_1.txt to out_dir_1/file_a.txt, and copies
  # in_dir_2/file_2.txt to out_dir_2/file_b.txt.
  - paths: ['in_dir_1/file_1.txt', in_dir_2/file_2.txt']
    as: ['out_dir_1/file_a.txt', 'out_dir_2/file_b.txt']
```

Using `strip_prefix` and/or `add_prefix`:

```
  - paths: ['my/deep/nested/dir']
    strip_prefix: 'my/deep/nested'
    add_prefix: 'other_dir'
```

The function `safeRelPath` is also changed so that it no longer rejects
absolute paths, it just treats them as relative paths. That means it's
not a problem if the user forgets to remove leading slashes from their
paths. This was originally intended as a security measure to protect
against malicious spec files exfiltrating data from the user's machine,
but it's just as safe to convert the absolute path to a relative path.