Create dest dir if nonexistent, rather than failing
There are no new calls to mkdir in this PR because we already have code
that creates directories as needed. We just needed to remove a
validation check.

Previously, we enforced that the destination dir already exists. This
was intended to support the user journey where we're installing into the
root of a git repo. However, a new user journey has emerged where we're
installing a new directory into an existing git repo. In this case it's
annoying to pre-create the directory and/or hackily precede every path
in the spec file with a subdirectory name.