Add default log level and mode
The main point is that we want human-friendly log output by default, not
JSON output which is currently the default. By setting ABC_LOG_MODE=dev,
we accomplish that.

There was some related conversation about an abandoned alternative
approach in https://github.com/abcxyz/pkg/pull/146. We chose not to do
that, and did this PR instead.

Also add docs about logging, which had been overlooked.