Refactor copyRecursive to add callback
This PR lays the groundwork for the "destination includes" feature
coming in a following PR.

We add a callback to the copyRecursive function that lets the handling
of each file be specified as it's crawled (overwrite or skip). In the
following PR we'll add a new option to back up each file if it's being
overwritten. It also informs the caller of the names of each crawled
file.

Also, fix a bug found while testing this change: add the `O_TRUNC` option
when overwriting a file. When overwriting a longer file with a shorter
file, we were leaving some of the old file contents at the end of the
file.

Also some misc cleanups:

 - rename `from`->`src` and `to`->`dst` in render_action_test.go. We agreed on
   this renaming but missed some spots.

 - factor out `copyFile()` from the anonymous function inside copyRecursive(). It
   was getting too big, and we'll need it in the next PR when we want to
   copy files one at a time.

 - Add a `string_replacement` step to the spec file in TestRealRun to make
   the test more realistic.