Add default log level and mode (#83)
The main point is that we want human-friendly log output by default, not
JSON output which is currently the default. By setting ABC_LOG_MODE=dev,
we accomplish that. I tested locally and confirmed this.

There was some related conversation about an abandoned alternative
approach in https://github.com/abcxyz/pkg/pull/146. We chose not to do
that, and did this PR instead.

Also add docs about logging, which had been overlooked.

---------

Signed-off-by: Dave Revell <dave.revell@gmail.com>
Co-authored-by: Chen Shou <cshou@google.com>