Add flags to go-template context for print action (#98)
Fixes #85.

To be able to pass the entire set of flags around, I factored them out
of the Render struct into a new dedicated struct, `renderFlags`.

Q: Why only expose some flags? Why expose `--dest` but not
`--force-overwrite`?

A: API parsimony. We should limit the things we expose to just the
things which we know are useful, since we'll have to support this in the
future. We can always expose more flags later.

Q: Why only expose these flags to the print action?

A: We want to avoid the temptation for template authors to be
inappropriately clever. We think we've provided a single correct way to
do things, and we don't want people macking weird hacks unnecessarily.
We want template rendering to be deterministic, and not change in
surprising ways depending on flag values.

Since we're adding a new `{{.flags}}` namespace, we now forbid a
template from declaring an input named `flags`, to avoid collisions.