feat: `include` files from destination dir (#96)
With this change, spec files can now `include` a file from the
destination dir rather than the template dir. This lets an existing file
be modified; before, a template could only overwrite an existing file
without regard for its contents.

When a file is included from the destination, it goes on an allowlist of
files that are allowed to be overwritten. This stands in contrast to
default policy of refusing to overwrite existing files (unless
`--force-overwrite`). It would be pointless to take a file from the
destination and then complain about overwriting it later; replacing it
was the whole point of including.

Before a file is overwritten in the destination directory, the old
version of the file is backed up in `~/.abc/backups/$timestamp/`.