Add an API for running Render programmatically
This exposes the RenderFlags struct for external use, and adds the
Render.RunWithFlags method.

This is PR 1 of 3-ish that will add a test framework for templates. The
goal is to (1) allow template developers to test their templates and (2)
allow abc CLI developers to verify that code changes haven't broken
templates.

Reviewers, all of the interesting changes are in render.go, and the rest
is just changed capitalization and renaming.