Move test helpers to testutil/
This is change 2 of 3-ish for that will add a test framework for templates. The goal is to (1) allow template developers to test their templates and (2) allow abc CLI developers to verify that code changes haven't broken templates.

This change:

 - Makes the test helpers usable outside the "commands" package
 - Changes the names slightly for conciseness

The convention I chose is to import the testutil package as
"abctestutil" to avoid colliding with the github.com/abcxyz/pkg/testutil package.