Optionally prompt for template inputs
What is this: if the user supplies the `--prompt` flag, then the user
will be prompted for template inputs (except those already provided by
`--input`).

Why? Because this makes it easy to run a template without knowing its
inputs in advance, and can help avoid lengthy command lines. It also
makes for a good demo.

Why is it opt-in with `--prompt` rather than defaulting to on? Consider this motivating use case: there's a playbook abc command
that the user is copy-pasting, and this command has always worked
with the  --input values provided in the playbook. There is no need
for prompting. Then the upstream template developer adds a new input
that is not anticipated by the playbook. In this case, we'd rather
just fail outright than have the CLI prompt for the missing input. If
we prompted the user, they'd be tempted to enter something creative
rather than realize that their playbook needs to be updated.