Remove the `--spec` flag (#141)
This was a group decision at this week's team meeting. We think it's
cleaner to have a template be unambiguously identified by its
`<location>` string, rather than a (location,--spec) tuple. This
simplifies the mental model for users.

So, now we always look for `spec.yaml` in the root directory of the
template location that was provided by the user.