feat: add --debug-scratch-contents flag (#146)
This is a feature for template authors that makes it easier to
understand the state of the scratch directory.
If the flag is provided, we print a recursive directory listing of the
scratch directory after each step of the spec file.

This addresses part of #121. It was useful in debugging #145.