Add hierarchical variable scopes (#153)
This is one of several PRs adding the `for_each` feature (#127).

In `for_each`, we add the concept of an "key", which is the index
variable of the loop that takes on a sequence of values. This means that
we need a way to bring a new variable into scope for the lifetime of the
for_each loop.

The solution is to do lexical scoping like in a programming language.
The "body" of the for_each is its own scope, and the for_each key is
only in-scope for those actions. If there's a variable of the same name
in an outer scope, it is temporarily shadowed until the inner scope is
done.

Perhaps controversial: the `print` action has privileged access to some
flag values, by doing lookups like `{{.flags.dest}}`. This was a weird
outlier, because `flags` was the only thing we exposed to users that
wasn't a string (it was a map). In this PR we change the lookup syntax
to `{{.flag_dest}}` to preserve the conceptual uniformity of "all names
that can be looked up by the user return strings." This saves much
complexity. To go along with this change, we prevent the user from
declaring any inputs starting with `_`, which reserves all those names
for our own use.

Alternatives considered: we could have just kept everything as a map
instead of introducing the "scope" type. We could have just temporarily
added a new variable to the map while the for_each loop was in scope.
This option wasn't taken because (1) mutability causes problems and (2)
scopes serve as a better theoretical foundation for future needs.