Partial for_each support (#155)
This adds partial support for the `for_each` action (#127). This
includes the core logic for running a sequence of steps over each of a
list of values.

This only supports iterating over a literal YAML list of values. Future
PRs will add CEL integration to iterate over a programmatically defined
list.

Docs are also still missing.