feat: for_each over CEL expression
Previously, we could only iterate over a hardcoded YAML list of values.
This PR adds the `values_from: 'some_cel_expression'` field to
programmatically compute a list of strings to iterate over.

We may want to add caching of compiled CEL programs in the future to
avoid paying a few milliseconds per compilation. We don't currently
have any use cases where a for_each action will run often enough for
that to matter, but it's conceivable that the latency would be
noticeable for deeply-nested for_each actions.

Still to do: write docs and examples