Allow YAML decoding to change depending on apiVersion
This is a hack, but I think it's the best option we have. It's a hack
because it uses mutable global state where it really doesn't
theoretically need to, but we're limited by our libraries.

In this PR we create a global variable that stores the apiVersion of the
YAML document currently being parsed. We add a mutex to ensure that only
one document can be parsed at a time.

Alternatives considered:

 - threadlocals: don't exist in Go
   (https://github.com/golang/go/issues/18802)
 - use a different YAML parser library that plumbs a context.Context to
   UnmarshalYAML(): github.com/goccy/go-yaml almost worked, but it
   doesn't provide line/column position information, which is a dealbreaker.
 - Just skip it and don't provide schema versioning: the schema would
   turn into a mess over time.
 - Write our own YAML decoder, potentially reusing
   https://pkg.go.dev/github.com/goccy/go-yaml/parser: would be lots of
   work.

This is a prelude to #130.