Clean up positional error reporting
We had two separate similar mechanisms for reporting an error along with
the position in the spec.yaml where it happened. Both were overly
verbose. Now there's one simple, concise way.

As a bonus we get to remove lots of nolint annotations because functions
named Errorf() are exempted from the `wrapcheck` linter.