Run presubmits on all PRs (#170)
We'll run the lint&test actions for all PRs, not just those targeting
`main`.

This improves the "stacked PR" workflow where the user breaks a change
into several PRs, each one "based" on the previous PR, and sends them
separately for review. Before this change, only the first PR in the
stack would automatically be linted&tested.

If this works sanely, we can consider making a similar change to other
abcxyz repos.