Advise the user to update if api_version is unknown
We anticipate adding new features and YAML fields. When a template uses
these new fields but the user is using an old binary, then there will be
error about the api_version being unrecognized. This error might be
confusing, so now we'll also tell the user to go update their binary.