Suggest upgrading api_version when it would help (#231)
2nd attempt at this PR after messing up my git technique. The contents
are the same as last time (#230)

I'm trying to help users be less confused when they try to use a feature
that's not supported in the api_version that they're using. When a YAML
file fails parsing or validation, we'll suggest that the user upgrade to
the newest api_version, with one condition. We'll only make this
suggestion if the file actually does parse and validate successfully
under the newest api_version.

For example, suppose you try to use the new `if` feature in a YAML file
whose `api_version` is v1alpha`. The error message will say
(paraphrasing) "your YAML file isn't valid, but if you change the
api_version field to v1beta1 instead if will be valid."