Untangle tech debt in file separators in tests
There are a bunch of tests that deal with converting file path
separators, but this can all be cleaned way up by moving it down into
the helper functions that are already being used.

Also, make WriteAll* call t.Fatal instead of returning error, to be
consistent with LoadAll* and save boilerplate in tests.