Add parser and downloader for git-hosted templates (#242)
This is PR 2/N to replace go-getter with our own git downloader. In this
PR we:

 - resolve template locations like "github.com/myorg/myrepo@latest"
- attempt multiple ways of parsing a location, so we can try treating a
given template source as either git or local and keep the first one that
parses successfully

Anticipated comments:
Q: Why is gitSourceParser so parameterizable and indirect? It's a
singleton, it doesn't need to be so flexible.
A: I'm going to add another instance of gitSourceParser in a future PR
that will parse go-getter-style URLs like
`github.com/abcxyz/abc.git//t/react_template?ref=latest`. So it won't be
a singleton anymore.