Opt-opt TestParseSource from parallel testing
Previously we opted out the subtests individually from parallel testing,
but on further thought we actually need to opt-out the entire test.
Since we call os.Chdir() in this test, we need to guarantee that we're
not running at the same time as another test which might possibly also
call os.Chdir().