Implement "canonical template sources"
A canonical template source is one that isn't tied to the machine where
it's running. For example, a template source like
`github/abcxyz/abc/t/rest_server` is canonical because anyone can
reference it. A template source like "/home/dave/my_unpacked_tarball" is
not canonical because that directory only exists on my machine.