fix: the --prompt feature was broken recently
We recently rewrote the input handling logic to make the c.flags.Input
field immutable and update a different mutable map as input values were
processed. However, we missed a spot that was left working the old way.
This resulted in inputs being ignored in the case where they were interactively entered by the user.