Implement "canonical template sources" (#283)
A canonical template source is one that isn't tied to the machine where
it's running. For example, a template source like
`github/abcxyz/abc/t/rest_server` is canonical because anyone can
reference it. A template source like "/home/dave/my_unpacked_tarball" is
not canonical because that directory only exists on my machine. As a
complicated special case, a path like `/home/dave/git/myrepo` may or may
not be canonical, depending on whether the destination is in the same
git repo as the source. See the internal design doc (ping me for a
link).

Also, we add a new common function `IsStatNotExist()` to factor out the
common pattern of checking whether we Stat()ed a path that doesn't
exist.