chore: allow versions with branch names or SHA (#290)
abc CLI `v0.3.0` introduced a constraint to only allowing versions with
`latest` or `semver` tag. We no longer want to force the tagged semver
constraint on users so this change will allow versioning with branch
names, including `@main`, or long commit SHA's. Short commit SHA's are
not supported and will be tried as branch name or tag.

Resolves #288