feat: write manifest file if --manifest is provided (#304)
Manifest files are part of the upcoming upgrade feature. They capture
all the important information used to render the template, along with
file hashes so we can detect when users have modified the template's
output files.

This feature is still experimental and doesn't have a stable API, so
it's OK that we're modifying the YAML models without bumping
api_version.

The --manifest flag defaults to false so we don't change any user
experience yet.