feat: add template version to the manifest file (#327)
Users are likely to want to know what version of a template was
rendered, and this provides it. This allows things like (1) render a
template at the same version as an already-rendered template (2) grep
for renderings of templates at a specific version.

Most of the complexity here is in dealing with git to retrieve info
about tags and SHAs. We already have a git clone of the template repo
(usually), we just have to inspect it to find out what version number we
have.

There's a interesting design decision in here. To find the template
version, we'll examine the git workspace and the HEAD commit. If there's
a tag pointing to HEAD, then we'll use that as the version. If there are
multiple such tags, we'll take the most recent in semver order. If the
tags aren't semver, we tiebreak using reverse-alphabetical order. If
there are no tags, we'll just use the full SHA as the version number. We
never use branch names for versions because branches are too mutable.