Refactor rendering into a Render() function instead of a command
Template rendering has become a subroutine inside other commands. For
example, the `golden-test` command already runs rendering internally,
somewhat hackishly by constructing a render.Command{} struct. We're
about to something similar for the `upgrade` command. Therefore, now is
the time to clean this up and make rendering a library function that can
be cleanly invoked as part of any command.