Refactor rendering into a Render() function instead of a command (#345)
Template rendering has become a subroutine inside other commands. For
example, the `golden-test` command already runs rendering internally,
somewhat hackishly by constructing a render.Command{} struct. We're
about to do something similar for the `upgrade` command. Therefore, now
is the time to clean this up and make rendering a library function that
can be cleanly invoked as part of any command. This resolves a TODO in
golden-test that we should do exactly this.

This PR is essentially:
- Most of the contents of **commands**/render/render.go are moved to
**common**/render/render.go, and massaged a bit for clarity and
cleanliness
- The action_*.go files (like action_print.go) are likewise moved from
commands/render to common/render

Still to do: migrate the tests from commands/render/render_test.go to
common/render. As of this PR, all the tests for rendering are testing
the command, not the library function, which isn't right. But this PR is
already too big.