Give a more helpful error if the user gives an invalid template location
These are mistakes that the user is likely to make, and we want to print
a more helpful error message.

Mistake 1: the user provides the path to the spec.yaml as the template
path, where they're actually supposed to provide the path to the
directory containing the spec.yaml. We add a very clear error for
this
case.

Mistake 2: providing some other filename instead of a directory name: we provide a warning that filenames can't be parsed as local template paths.