Streamline the tracking of features enabled in a given api_version (#359)
Background: we have to keep the semantics of a given api_version stable
even though we're constantly adding and changing features. We have a
`Features` struct on the Spec model that tracks which features are
enabled based on the api_version that's declared in the spec.yaml.

This PR:
- Uses a single Features struct rather than creating separate Features
structs per schema version. That was unnecessarily verbose and complex.
- Fixes a bug in the v1beta2->v1beta3 upgrader that disregarded the
value of OmitGlobs when the spec.yaml declared api_version=v1beta1.
- Uses a non-pointer struct rather than a pointer struct. This saves a
bunch of "initalize pointer to point to empty struct" code, especially
in tests.