Add builtin variables for _git_sha, _git_short_sha, and _git_tag (#360)
As seen in the internal design doc, these variables will be populated
when the template is being rendered from git.

The motivating use case is for terraform module pinning. We sometimes
want to render a template that pins a terraform module to the same SHA
that the template is rendered it.