Add location_type to the upgrade manifest YAML model
There's a design decision here. We're treating template location strings differently in manifest files
than when we parse template locations provided by users on the command
line.

 - Template locations on the command line don't explicitly specify the
   template location type. We just try to parse it in a few different
   ways and see what works (i.e. is this a github URL or a local
   directory?)
 - In manifest files, we explicitly record the template location type
   (either "local_git" or "remote_git"), and *don't* try to autodetect
   anything.

Why? There are two important differences:

 - Manifest files are written and read only by programs, not people. We
   can pay a little cost in verbosity to avoid the ambiguity of trying
   to guess whether a location is a URL or a local file.
 - Manifest files store template locations without a version string
   (i.e. `@1.2.3`), so we can't just reuse the same code anyway.

How this will be used in a future PR: we'll `switch` based on the value
of `location_type` and construct the appropriate downloader struct,
either `gitDownloader` or `localDownloader`. We'll pass the
`template_location` string into the downloader, and each downloader will
interpret it differently.