Make builtin vars only in scope if overridden (#371)
To help golden-test authors write good tests, we want to force them to
provide fake values for all the builtin vars that they use, without
accidentally relying on a default value of empty string. By forcing the
test authors to think through what they're testing, they'll be more
likely to cover both the empty-string and non-empty-string cases.

To this, if a template is being run in a golden-test, and the template
references a builtin var, and that var doesn't have an override defined
in test.yaml's `builtin_vars` field, the test will fail with a message
about an unknown variable.

A future PR will make the error easier to understand and obvious how to
fix the test.