fix: template with no output should create an empty output dir (#384)
The former behavior was that the output directory would not be created
in the edge case where the template produced no output files. This was
problematic for two reasons:

- It broke `golden-test record`; it expected to find an output directory
 - Empty output is semantically valid. Suppose there's a template that
   has a bunch of optional steps that are all skipped.

This requires changes to template rendering, golden-test recording, and
golden-test verification.

I also decided to use the `.abc` directory name for both manifests and
golden-test recording metadata. It's simple and consistent to have just
one metadata directory name that we use everywhere.