Workaround for git mangling line-endings in golden tests (#401)
This is a hacky workaround for #400, but doesn't fully address that.

By running `git config --global core.autocrlf false` in the CI workflow,
we prevent git from mangling the line endings when checking out the
golden test recorded output. See #400 for a more thorough writeup of the
problem.

We also add a test for the `git_vars` example template, which exercises
the fix.