Add the concept of unreleased api_versions (#430)
Currently, it can be hard to release a new binary version because any
unfinished features and api_versions will be exposed in that release.
For example, we can't release a new binary version if api_version
`v1beta4` is unfinished, because the new binary will put `api_version:
v1beta4` in all the yaml files that it creates.

Instead, we want the ability to release a new binary version at any
time, with unfinished features hidden behind an api_version that is
still hidden from the public.

To do this, we mark each release struct with a boolean indicating
whether it is released or not. api_versions that are under construction
will not be used by official release binaries. When a release build of
abc creates a yaml file, the header will use
api_version=$last_officially_released_api_version, not the api_version
that's under construction at the time of release.