Create api_version v1beta5, same as v1beta4 (#431)
This is a simple hack for a very complex reason.

In #428, there's a problem where abc CLI version 0.5.0 *thinks* that it
supports api_version v1beta4, but it really doesn't. It just supports
the in-progress subset of v1beta4 that was implemented at the time that
0.5.0 was released. This causes a problem because when abc v0.5.0 sees a
test.yaml that has api_version v1beta4, it *should* say "I don't
understand this version, please upgrade your abc CLI" but what it will
actually say is a weird test failure involving the `.gitkeep` file.

By bumping the soon-to-be-released api_version to v1beta5, that will be
beyond what abc v0.5.0 thinks it can support, so it will correctly ask
the user to upgrade.

The underlying root cause (abc CLI thinks it supports api_versions that
it really doesn't) is fixed in #430 , but obviously we can't fix v0.5.0
which is already out in the wild.