Add the ability to force a given manifest filename (#463)
Background: during a normal render operation, when manifests are
enabled, a manifest is generated in the `.abc` subdirectory of the
destination, and this manifest file has an unpredictable filename
containing a timestamp and some other stuff.

When doing a template *upgrade* (not just a normal render operation),
the upgrade operation must replace the existing manifest with a new
manifest. In order for this to happen, the new manifest filename must be
the same as the old manifest filename. Therefore, the Render operation
now accepts an optional parameter to force the output manifest file to
have a certain name. This optional parameter is only intended to be used
by the template upgrade process.

This functionality will be used in my next PR which starts implementing
the core upgrade logic.