Add the concept of "ultimate" destination to rendering logic (#467)
Before now, the Render() function passed through the output directory to
the Download() function, so the downloader could determine if the
template source location was canonical (input and output are in the same
git workspace). But now we need more info, because of the subleties of
upgrading.

When doing `abc templates upgrade`, we'll use the Render() function as a
building block in a way that's a little different than how it's used by
`abc templates render`. The `abc templates upgrade` doesn't use the same
directory for the render output as for the final upgrade destination,
which means we need to now provide both directories to Render(). Hence
the addition of the DestDirUltimate field.